RegressionCNN/requirements.txt
albumentations==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==3.0.1
bleach==4.1.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
certifi==2021.10.8
colorama==0.4.4
configparser==5.0.2
cycler==0.10.0
datatest==0.11.1
defusedxml==0.7.1
Deprecated==1.2.11
docopt==0.6.2
entrypoints==0.3
eulerangles==1.0.1
executing==0.6.0
ExtensionClass==4.5.0
fsspec==0.8.5
future==0.18.2
google-pasta==0.2.0
graphviz==0.16
hiddenlayer==0.3
humanize==3.2.0
icecream==2.1.0
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.4.0
importlib-resources==5.4.0
imread==0.7.4
imutils==0.5.4
install==1.3.4
ipywidgets==7.6.5
jedi==0.17.0
Jinja2==3.0.3
joblib==1.0.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-console==6.4.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras==2.4.3
linalg==1.0.4
MarkupSafe==2.0.1
mathutils==2.81.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mpmath==1.2.1
munch==2.5.0
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netron==4.8.1
networkx==2.5
nibabel==3.2.1
nilearn==0.7.0
nose==1.3.7
notebook==6.4.7
nvidia-ml-py==11.450.51
nvitop==0.5.2.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.5.0
Pillow==8.3.2
pipreqs==0.4.11
plyfile==0.7.3
pretrainedmodels==0.7.4
prometheus-client==0.12.0
protobuf==3.14.0
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyrsistent==0.18.1
pytorch-lightning==1.1.8
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.2.2
QtPy==2.0.0
Record==3.5
requests-oauthlib==1.3.0
scikit-image==0.18.1
scikit-learn==0.24.1
Send2Trash==1.8.0
Shapely==1.7.1
SimpleITK==1.2.4
sklearn==0.0
sympy==1.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.1.0
tifffile==2021.2.1
torch==1.8.0
torchaudio==0.8.0a0+a751e1d
torchio==0.18.29
torchsummary==1.5.1
torchvision==0.9.0
torchviz==0.0.1
tqdm==4.56.2
transformations==2020.1.1
transforms3d==0.3.1
webencodings==0.5.1
widgetsnbextension==3.5.2
wrapt==1.12.1
yarg==0.1.9
yarl==1.5.1
pyscreenshot==3.0
nvitop
numpy
oyaml
matplotlib
